4.4.0-0.0.1: [fix]: Default Export (#196)
* Fixes export issues and updates multichain docs

* Change to import type only
