+++ src/multichain/index.ts
@@ -1,5 +1,5 @@
 import { Observable, Subject } from 'rxjs'
-import Blocknative from '../sdk'
+import type SDK from '../'
 import subscribe from './subscribe'
 import unsubscribe from './unsubscribe'
 
@@ -14,15 +14,16 @@ import {
 class MultiChain {
   public apiKey: string
   public ws: WebSocket | void
-  public connections: Record<ChainId, Blocknative | null>
+  public connections: Record<ChainId, SDK | null>
   public transactions$: Observable<EthereumTransactionData>
   public errors$: Subject<SDKError>
   public subscribe: typeof subscribe
   public unsubscribe: typeof unsubscribe
+  public Blocknative: typeof SDK
 
   protected onTransaction$: Subject<EthereumTransactionData>
 
-  constructor(options: MultiChainOptions) {
+  constructor(options: MultiChainOptions, Blocknative: typeof SDK) {
     const { apiKey, ws } = options
 
     this.apiKey = apiKey
@@ -31,6 +32,7 @@ class MultiChain {
     this.onTransaction$ = new Subject()
     this.transactions$ = this.onTransaction$.asObservable()
     this.errors$ = new Subject()
+    this.Blocknative = Blocknative
     this.subscribe = subscribe.bind(this)
     this.unsubscribe = unsubscribe.bind(this)
   }
